Move component correct place
