--- src/stories/components/Icon/IconNames.ts
@@ -1,11 +1,0 @@
-import ArrowLink from '../svg/ArrowLink';
-import AddIcon from '../svg/add';
-import { LinkIcon } from '../svg/link-icon';
-
-export const icons = {
-  add: AddIcon,
-  arrowLink: ArrowLink,
-  linkIcon: LinkIcon,
-};
-
-export type IconNames = keyof typeof icons;
