+++ src/stories/containers/RecognizedDelegates/components/Icon/Icon.tsx
